Obtain a Geocorr22 Crosswalk

Query Geocorr22 (https://mcdc.missouri.edu/applications/geocorr2022.html) for a crosswalk between two geographies for all 51 states and Puerto Rico.

Usage

get_geocorr_crosswalk(
  source_geography,
  target_geography,
  weight = c("population", "housing", "land"),
  cache = NULL,
  overwrite_cache = FALSE
)

Arguments

source_geography

Source geography type. Supported geographies, at present, are: c("place", "county", "tract", "blockgroup", "zcta", "puma22", "cd119", "cd118"). Note "cd" stands for "congressional district".

target_geography

Target geography type. See source_geography.

weight

Weighting variable. One of c("population", "housing", "land").

cache

NULL by default. If not NULL, the path at which to cache the results.

overwrite_cache

FALSE by default. If TRUE, the existing cached crosswalk will be replaced.

Value

A dataframe representing the requested Geocorr22 crosswalk for all 51 states and Puerto Rico. Depending on the desired geographies, some fields may not be included.

state_fips

A two-digit identified for the state (or DC/PR), if applicable

state_abbreviation

A two-character abbreviation for the state (or DC/PR), if applicable

source_geoid

A unique identifier for the source geography

target_geoid

A unique identifier for the target geography

source_geography_name

The name of the source geography

target_geography_name

The name of the target geography

allocation_factor_source_to_target

The weight to interpolate values from the source geography to the target geography

allocation_factor_target_to_source

The weight to interpolate values from the source geography to the target geography

population_2020

The estimated overlap in population, if applicable

housing_2020

The estimated overlap in housing units, if applicable

land_area_sqmi

The overlap in land area, if applicable

weighting_factor

The attribute used to calculate allocation factors (one of population, housing, land)

Details

Note: this function is under development but does not yet support all of the geographies supported by Geocorr. Currently this includes: c("place", "county", "tract", "blockgroup", "zcta", "puma22", "cd119", "cd118").