Get NHGIS Geographic Crosswalk

Retrieves a geographic crosswalk from the IPUMS NHGIS API based on user-specified source and target geographies and years. Accepts various spellings and formats for geography names.

Usage

get_nhgis_crosswalk(
  source_year,
  source_geography,
  target_year,
  target_geography,
  api_key = NULL
)

Arguments

source_year

Character or numeric. Year of the source geography one of c(1990, 2000, 2010, 2020).

source_geography

Character. Source geography name. One of c("block", "block group part", "tract").

target_year

Character or numeric. Year of the target geography, one of c(1990, 2000, 2010, 2020).

target_geography

Character. Target geography name. One of c("block", "block group part", "tract", "county").

api_key

Character. NULL by default, in which case the function looks for an IPUMS_API_KEY environment variable.

Value

A data frame containing the crosswalk between the specified geographies.